database/requirements.txt
hvac
minio
pandas
python-dotenv
psycopg2-binary
confluent-kafka
